Image fitting - cleanup of code
Significant cleanup of code to allow for expansion of image cleaning routines
